feat: remove Old Settings
