+++ p2p_wallet/Scenes/DebugMenu/ViewModel/DebugMenuViewModel.swift
@@ -102,7 +102,6 @@ extension DebugMenuViewModel {
 extension DebugMenuViewModel {
     enum Menu: Int, CaseIterable {
         case solanaNegativeStatus
-        case newSettings
         case onboardingUsernameEnabled
         case onboardingUsernameButtonSkipEnabled
         
@@ -117,7 +116,6 @@ extension DebugMenuViewModel {
         var title: String {
             switch self {
             case .solanaNegativeStatus: return "Solana Negative Status"
-            case .newSettings: return "New Settings"
             case .onboardingUsernameEnabled: return "Onboarding Username"
             case .onboardingUsernameButtonSkipEnabled: return "Onboarding Username Skip Button"
             case .mockedApiGateway: return "[Onboarding] API Gateway Mock"
@@ -132,7 +130,6 @@ extension DebugMenuViewModel {
         var feature: Feature {
             switch self {
             case .solanaNegativeStatus: return .solanaNegativeStatus
-            case .newSettings: return .settingsFeature
             case .onboardingUsernameEnabled: return .onboardingUsernameEnabled
             case .onboardingUsernameButtonSkipEnabled: return .onboardingUsernameButtonSkipEnabled
             case .mockedApiGateway: return .mockedApiGateway
