+++ p2p_wallet/Scenes/Main/NewSettings/Subscenes/SeedPhraseDetail/SeedPhraseDetailView.swift
