--- p2p_wallet/Scenes/Main/Settings/Subscenes/Settings.SelectLanguageViewController.swift
@@ -1,57 +1,0 @@
-//
-//  Settings.SelectLanguageViewController.swift
-//  p2p_wallet
-//
-//  Created by Chung Tran on 12/10/2021.
-//
-
-extension Settings {
-    class SelectLanguageViewController: SingleSelectionViewController<LocalizedLanguage> {
-        override init(viewModel: SettingsViewModelType) {
-            super.init(viewModel: viewModel)
-
-            setSelectableLanguages()
-        }
-
-        override func setUp() {
-            super.setUp()
-            navigationItem.title = L10n.language
-        }
-
-        override func createCell(item: LocalizedLanguage) -> Cell<LocalizedLanguage> {
-            let cell = super.createCell(item: item)
-            cell.label.text = item.originalName?.uppercaseFirst
-            return cell
-        }
-
-        override func itemDidSelect(at index: Int) {
-            let originalSelectedItem = selectedItem
-            super.itemDidSelect(at: index)
-            showAlert(
-                title: L10n.switchLanguage,
-                message: L10n.doYouReallyWantToSwitchTo + " " + selectedItem?.localizedName?.uppercaseFirst + "?",
-                buttonTitles: [L10n.ok, L10n.cancel],
-                highlightedButtonIndex: 0
-            ) { [weak self] index in
-                guard index == 0, let language = self?.selectedItem
-                else {
-                    self?.reverseSelectedItem(originalSelectedItem: originalSelectedItem)
-                    return
-                }
-                self?.viewModel.setLanguage(language)
-            }
-        }
-
-        private func reverseSelectedItem(originalSelectedItem: LocalizedLanguage?) {
-            guard
-                let item = originalSelectedItem,
-                let index = (data.firstIndex { $0.item == item })
-            else { return }
-            super.itemDidSelect(at: index)
-        }
-
-        private func setSelectableLanguages() {
-            data = viewModel.selectableLanguages
-        }
-    }
-}
