--- p2p_wallet/Scenes/Main/Settings/Subscenes/Settings.UsernameViewController.swift
@@ -1,66 +1,0 @@
-//
-//  Settings.UsernameViewController.swift
-//  p2p_wallet
-//
-//  Created by Chung Tran on 12/10/2021.
-//
-
-import Foundation
-
-extension Settings {
-    class NewUsernameViewController: p2p_wallet.BaseViewController {
-        let viewModel: SettingsViewModelType
-
-        init(viewModel: SettingsViewModelType) {
-            self.viewModel = viewModel
-            super.init()
-            navigationItem.title = L10n.yourUsername
-
-            viewModel.navigationDriver
-                .drive(onNext: { [weak self] in self?.navigate(to: $0) })
-                .disposed(by: disposeBag)
-        }
-
-        override func build() -> UIView {
-            UIStackView(axis: .vertical, alignment: .fill) {
-                BEScrollView(contentInsets: .init(x: 0, y: 8), spacing: 18) {
-                    ReceiveToken.QrCodeCard(
-                        username: viewModel.getUsername(),
-                        pubKey: viewModel.getUserAddress(),
-                        token: .nativeSolana,
-                        showCoinLogo: false
-                    )
-                        .onCopy { [weak self] _ in
-                            self?.viewModel.copyUsernameToClipboard()
-                        }.onShare { [weak self] image in
-                            self?.viewModel.share(image: image)
-                        }.onSave { [weak self] image in
-                            self?.viewModel.saveImage(image: image)
-                        }
-
-                    UIView.greyBannerView {
-                        UILabel(
-                            text:
-                            L10n
-                                .yourUsernameAllowsYouToReceiveAnyTokenWithinTheSolanaNetworkEvenIfItIsNotIncludedInYourWalletTokenList,
-                            textSize: 15,
-                            numberOfLines: 0
-                        )
-                    }
-
-                }.padding(.init(x: 18, y: 0))
-            }
-        }
-
-        private func navigate(to scene: NavigatableScene?) {
-            guard let scene = scene else { return }
-            switch scene {
-            case let .share(item):
-                let vc = UIActivityViewController(activityItems: [item], applicationActivities: nil)
-                present(vc, animated: true, completion: nil)
-            default:
-                break
-            }
-        }
-    }
-}
