+++ p2p_wallet/Services/FeatureFlags/Features.swift
@@ -9,7 +9,6 @@ public extension Feature {
     static let coinGeckoPriceProvider = Feature(rawValue: "coinGeckoPriceProvider")
     static let sellScenarioEnabled = Feature(rawValue: "keyapp_sell_scenario_enabled")
     static let buyBankTransferEnabled = Feature(rawValue: "buy_bank_transfer_enabled")
-    static let settingsFeature = Feature(rawValue: "settingsFeature")
     static let newOnboardingFlow = Feature(rawValue: "newOnboardingFlow")
 
     // Username
